Fcf

Description

First-class type families

For example, here is a regular type family:

type family   FromMaybe (a :: k) (m :: Maybe k) :: k
type instance FromMaybe a 'Nothing  = a
type instance FromMaybe a ('Just b) = b

With Fcf, it translates to a data declaration:

data FromMaybe :: k -> Maybe k -> Exp k
type instance Eval (FromMaybe a 'Nothing)  = a
type instance Eval (FromMaybe a ('Just b)) = b

Fcfs can be higher-order.

The kind constructor Exp is a monad: there's (=<<) and Pure.

Essential language extensions for Fcf:

{-# LANGUAGE
    DataKinds,
    PolyKinds,
    TypeFamilies,
    TypeOperators,
    UndecidableInstances #-}

First-class type families

type Exp a = a -> Type

Kind of type-level expressions indexed by their result type.

type family Eval (e :: Exp a) :: a

Expression evaluator.

type (@@) (f :: k1 -> Exp k) (x :: k1) = Eval (f x)

Apply and evaluate a unary type function.

Functional combinators

data Pure (b :: a) (c :: a)

data Pure1 (c :: a -> b) (d :: a) (e :: b)

data Pure2 (d :: a -> b -> c) (e :: a) (f :: b) (g :: c)

data Pure3 (e :: a -> b -> c -> d) (f :: a) (g :: b) (h :: c) (i :: d)

data Pure4 (f :: a -> b -> c -> d -> e) (g :: a) (h :: b) (i :: c) (j :: d) (k :: e)

data Pure5 (g :: a -> b -> c -> d -> e -> f) (h :: a) (i :: b) (j :: c) (k :: d) (l :: e) (m :: f)

data Pure6 (h :: a -> b -> c -> d -> e -> f -> g) (i :: a) (j :: b) (k :: c) (l :: d) (m :: e) (n :: f) (o :: g)

data Pure7 (i :: a -> b -> c -> d -> e -> f -> g -> h) (j :: a) (k :: b) (l :: c) (m :: d) (n :: e) (o :: f) (p :: g) (q :: h)

data Pure8 (j :: a -> b -> c -> d -> e -> f -> g -> h -> i) (k :: a) (l :: b) (m :: c) (n :: d) (o :: e) (p :: f) (q :: g) (r :: h) (s :: i)

data Pure9 (k :: a -> b -> c -> d -> e -> f -> g -> h -> i -> j) (l :: a) (m :: b) (n :: c) (o :: d) (p :: e) (q :: f) (r :: g) (s :: h) (t :: i) (u :: j)

data ((c :: a -> Exp b) =<< (d :: Exp a)) (e :: b) infixr 1

data ((c :: Exp a) >>= (d :: a -> Exp b)) (e :: b) infixl 1

data ((d :: b -> Exp c) <=< (e :: a -> Exp b)) (f :: a) (g :: c) infixr 1

type LiftM = (=<<) :: (a -> Exp b) -> Exp a -> b -> Type

data LiftM2 (d :: a -> b -> Exp c) (e :: Exp a) (f :: Exp b) (g :: c)

data LiftM3 (e :: a -> b -> c -> Exp d) (f :: Exp a) (g :: Exp b) (h :: Exp c) (i :: d)

data Join (b :: Exp (Exp a)) (c :: a)

data ((c :: a -> b) <$> (d :: Exp a)) (e :: b) infixl 4

data ((c :: Exp (a -> b)) <*> (d :: Exp a)) (e :: b) infixl 4

data Flip (d :: a -> b -> Exp c) (e :: b) (f :: a) (g :: c)

data ConstFn (c :: a) (d :: b) (e :: a)

data ((c :: a -> Exp b) $ (d :: a)) (e :: b) infixr 0

Note that this denotes the identity function, so ($) f can usually be replaced with f.

Operations on common types

Pairs

data Uncurry (d :: a -> b -> Exp c) (e :: (a, b)) (f :: c)

data Fst (c :: (a, b)) (d :: a)

data Snd (c :: (a, b)) (d :: b)

data ((a :: b -> Exp c) *** (d :: b' -> Exp c')) (e :: (b, b')) (f :: (c, c')) infixr 3

Specialization of Bimap for pairs.

Either

data UnEither (d :: a -> Exp c) (e :: b -> Exp c) (f :: Either a b) (g :: c)

data IsLeft (c :: Either a b) (d :: Bool)

data IsRight (c :: Either a b) (d :: Bool)

Maybe

data UnMaybe (c :: Exp b) (d :: a -> Exp b) (e :: Maybe a) (f :: b)

data FromMaybe (a :: k) (b :: Maybe k) (c :: k)

data IsNothing (b :: Maybe a) (c :: Bool)

data IsJust (b :: Maybe a) (c :: Bool)

Lists

data ((b :: [a]) ++ (c :: [a])) (d :: [a])

List catenation.

Example

Expand

>>> data Example where Ex :: a -> Example  -- Hide the type of examples to avoid brittleness in different GHC versions
>>> :kind! Ex (Eval ([1, 2] ++ [3, 4]) :: [Natural])
Ex (Eval ([1, 2] ++ [3, 4]) :: [Natural]) :: Example
= Ex [1, 2, 3, 4]

data Head (b :: [a]) (c :: Maybe a)

data Last (b :: [a]) (c :: Maybe a)

data Tail (b :: [a]) (c :: Maybe [a])

data Cons (b :: a) (c :: [a]) (d :: [a])

Append an element to a list.

Example

Expand

>>> :kind! Eval (Cons 1 [2, 3])
Eval (Cons 1 [2, 3]) :: [Natural]
= [1, 2, 3]
>>> :kind! Eval (Cons Int [Char, Maybe Double])
Eval (Cons Int [Char, Maybe Double]) :: [*]
= [Int, Char, Maybe Double]

data Snoc (b :: [a]) (c :: a) (d :: [a])

Append an element to the end of a list.

Example

Expand

>>> :kind! Eval (Snoc [1,2,3] 4)
Eval (Snoc [1,2,3] 4) :: [Natural]
= [1, 2, 3, 4]

data Cons2 (c :: (a, b)) (d :: ([a], [b])) (e :: ([a], [b]))

Append elements to two lists. Used in the definition of Unzip.

data Init (b :: [a]) (c :: Maybe [a])

data Uncons (b :: [a]) (c :: Maybe (a, [a]))

data Unsnoc (b :: [a]) (c :: Maybe ([a], a))

Decompose a list into init and last

Example

Expand

>>> :kind! Eval (Unsnoc '[])
Eval (Unsnoc '[]) :: Maybe ([a], a)
= Nothing

>>> :kind! Eval (Unsnoc '[1])
Eval (Unsnoc '[1]) :: Maybe ([Natural], Natural)
= Just '( '[], 1)

>>> :kind! Eval (Unsnoc '[1,2,3])
Eval (Unsnoc '[1,2,3]) :: Maybe ([Natural], Natural)
= Just '([1, 2], 3)

data Singleton (b :: a) (c :: [a])

data Null (b :: [a]) (c :: Bool)

data Length (b :: [a]) (c :: Nat)

data Reverse (b :: [a]) (c :: [a])

Reverse a list.

Example

Expand

>>> :kind! Eval (Reverse [1,2,3,4,5])
Eval (Reverse [1,2,3,4,5]) :: [Natural]
= [5, 4, 3, 2, 1]

data Intersperse (b :: a) (c :: [a]) (d :: [a])

Intersperse a separator between elements of a list.

Example

Expand

>>> :kind! Eval (Intersperse 0 [1,2,3,4])
Eval (Intersperse 0 [1,2,3,4]) :: [Natural]
= [1, 0, 2, 0, 3, 0, 4]

data Intercalate (b :: [a]) (c :: [[a]]) (d :: [a])

Join a list of words separated by some word.

Example

Expand

>>> :kind! Eval (Intercalate '[", "] [ '["Lorem"], '["ipsum"], '["dolor"] ])
Eval (Intercalate '[", "] [ '["Lorem"], '["ipsum"], '["dolor"] ]) :: [TL.Symbol]
= ["Lorem", ", ", "ipsum", ", ", "dolor"]

data Foldr (c :: a -> b -> Exp b) (d :: b) (e :: t a) (f :: b)

Right fold.

Example

Expand

>>> :kind! Eval (Foldr (+) 0 [1, 2, 3, 4])
Eval (Foldr (+) 0 [1, 2, 3, 4]) :: Natural
= 10

data UnList (c :: b) (d :: a -> b -> Exp b) (e :: [a]) (f :: b)

This is Foldr with its argument flipped.

data Concat (a :: t m) (b :: m)

Concatenate a collection of elements from a monoid.

Example

Expand

For example, fold a list of lists.

Concat :: [[a]] -> Exp [a]

>>> :kind! Eval (Concat ([[1,2], [3,4], [5,6]]))
Eval (Concat ([[1,2], [3,4], [5,6]])) :: [Natural]
= [1, 2, 3, 4, 5, 6]
>>> :kind! Eval (Concat ([[Int, Maybe Int], [Maybe String, Either Double Int]]))
Eval (Concat ([[Int, Maybe Int], [Maybe String, Either Double Int]])) :: [*]
= [Int, Maybe Int, Maybe [Char], Either Double Int]

data ConcatMap (c :: a -> Exp [b]) (d :: t a) (e :: [b])

Map a function and concatenate the results.

This is FoldMap specialized to the list monoid.

data Unfoldr (c :: b -> Exp (Maybe (a, b))) (d :: b) (e :: [a])

Unfold a generator into a list.

Example

Expand

>>> data ToThree :: Nat -> Exp (Maybe (Nat, Nat))
>>> :{
type instance Eval (ToThree b) =
  If (4 TL.<=? b)
    Nothing
    (Just '(b, b TL.+ 1))
:}

>>> :kind! Eval (Unfoldr ToThree 0)
Eval (Unfoldr ToThree 0) :: [Natural]
= [0, 1, 2, 3]

See also the definition of Replicate.

data Replicate (b :: Nat) (c :: a) (d :: [a])

Repeat the same element in a list.

Example

Expand

>>> :kind! Eval (Replicate 4 '("ok", 2))
Eval (Replicate 4 '("ok", 2)) :: [(TL.Symbol, Natural)]
= ['("ok", 2), '("ok", 2), '("ok", 2), '("ok", 2)]

data Take (b :: Nat) (c :: [a]) (d :: [a])

Take a prefix of fixed length.

Example

Expand

>>> :kind! Eval (Take 2 [1,2,3,4,5])
Eval (Take 2 [1,2,3,4,5]) :: [Natural]
= [1, 2]

data Drop (b :: Nat) (c :: [a]) (d :: [a])

Drop a prefix of fixed length, evaluate to the remaining suffix.

Example

Expand

>>> :kind! Eval (Drop 2 [1,2,3,4,5])
Eval (Drop 2 [1,2,3,4,5]) :: [Natural]
= [3, 4, 5]

data SplitAt (b :: Nat) (c :: [a]) (d :: ([a], [a]))

Return a tuple where first element is xs prefix of length n and second element is the remainder of the list.

Example

Expand

>>> :kind! Eval (SplitAt 3 '[1,2,3,4,5])
Eval (SplitAt 3 '[1,2,3,4,5]) :: ([Natural], [Natural])
= '([1, 2, 3], [4, 5])

data TakeWhile (b :: a -> Exp Bool) (c :: [a]) (d :: [a])

Take the longest prefix of elements satisfying a predicate.

Example

Expand

>>> :kind! Eval (TakeWhile ((>=) 3) [1, 2, 3, 4, 5])
Eval (TakeWhile ((>=) 3) [1, 2, 3, 4, 5]) :: [Natural]
= [1, 2, 3]

data DropWhile (b :: a -> Exp Bool) (c :: [a]) (d :: [a])

Drop the longest prefix of elements satisfying a predicate, evaluate to the remaining suffix.

Example

Expand

:kind! Eval (DropWhile ((>=) 3) [1, 2, 3, 4, 5]) Eval (DropWhile ((>=) 3) [1, 2, 3, 4, 5]) :: [Natural] = [4, 5]

data Span (b :: a -> Exp Bool) (c :: [a]) (d :: ([a], [a]))

Span, applied to a predicate p and a list xs, returns a tuple: the first component is the longest prefix (possibly empty) of xs whose elements satisfy p; the second component is the remainder of the list.

See also TakeWhile, DropWhile, and Break.

Example

Expand

>>> :kind! Eval (Span (Flip (<) 3) [1,2,3,4,1,2])
Eval (Span (Flip (<) 3) [1,2,3,4,1,2]) :: ([Natural], [Natural])
= '([1, 2], [3, 4, 1, 2])

>>> :kind! Eval (Span (Flip (<) 9) [1,2,3])
Eval (Span (Flip (<) 9) [1,2,3]) :: ([Natural], [Natural])
= '([1, 2, 3], '[])

>>> :kind! Eval (Span (Flip (<) 0) [1,2,3])
Eval (Span (Flip (<) 0) [1,2,3]) :: ([Natural], [Natural])
= '( '[], [1, 2, 3])

data Break (b :: a -> Exp Bool) (c :: [a]) (d :: ([a], [a]))

Break, applied to a predicate p and a list xs, returns a tuple: the first component is the longest prefix (possibly empty) of xs whose elements do not satisfy p; the second component is the remainder of the list.

Example

Expand

>>> :kind! Eval (Break (Flip (>) 3) [1,2,3,4,1,2])
Eval (Break (Flip (>) 3) [1,2,3,4,1,2]) :: ([Natural], [Natural])
= '([1, 2, 3], [4, 1, 2])

>>> :kind! Eval (Break (Flip (<) 9) [1,2,3])
Eval (Break (Flip (<) 9) [1,2,3]) :: ([Natural], [Natural])
= '( '[], [1, 2, 3])

>>> :kind! Eval (Break (Flip (>) 9) [1,2,3])
Eval (Break (Flip (>) 9) [1,2,3]) :: ([Natural], [Natural])
= '([1, 2, 3], '[])

data Tails (b :: [a]) (c :: [[a]])

List of suffixes of a list.

Example

Expand

>>> :kind! Eval (Tails [0,1,2,3])
Eval (Tails [0,1,2,3]) :: [[Natural]]
= [[0, 1, 2, 3], [1, 2, 3], [2, 3], '[3]]

data IsPrefixOf (b :: [a]) (c :: [a]) (d :: Bool)

Return True when the first list is a prefix of the second.

Example

Expand

>>> :kind! Eval ([0,1,2] `IsPrefixOf` [0,1,2,3,4,5])
Eval ([0,1,2] `IsPrefixOf` [0,1,2,3,4,5]) :: Bool
= True

>>> :kind! Eval ([0,1,2] `IsPrefixOf` [0,1,3,2,4,5])
Eval ([0,1,2] `IsPrefixOf` [0,1,3,2,4,5]) :: Bool
= False

>>> :kind! Eval ('[] `IsPrefixOf` [0,1,3,2,4,5])
Eval ('[] `IsPrefixOf` [0,1,3,2,4,5]) :: Bool
= True

>>> :kind! Eval ([0,1,3,2,4,5] `IsPrefixOf` '[])
Eval ([0,1,3,2,4,5] `IsPrefixOf` '[]) :: Bool
= False

data IsSuffixOf (b :: [a]) (c :: [a]) (d :: Bool)

Return True when the first list is a suffix of the second.

Example

Expand

>>> :kind! Eval (IsSuffixOf [3,4,5] [0,1,2,3,4,5])
Eval (IsSuffixOf [3,4,5] [0,1,2,3,4,5]) :: Bool
= True

>>> :kind! Eval (IsSuffixOf [3,4,5] [0,1,3,2,4,5])
Eval (IsSuffixOf [3,4,5] [0,1,3,2,4,5]) :: Bool
= False

>>> :kind! Eval (IsSuffixOf '[] [0,1,3,2,4,5])
Eval (IsSuffixOf '[] [0,1,3,2,4,5]) :: Bool
= True

>>> :kind! Eval (IsSuffixOf [0,1,3,2,4,5] '[])
Eval (IsSuffixOf [0,1,3,2,4,5] '[]) :: Bool
= False

data IsInfixOf (b :: [a]) (c :: [a]) (d :: Bool)

Return True when the first list is contained within the second.

Example

Expand

>>> :kind! Eval (IsInfixOf [2,3,4] [0,1,2,3,4,5,6])
Eval (IsInfixOf [2,3,4] [0,1,2,3,4,5,6]) :: Bool
= True

>>> :kind! Eval (IsInfixOf [2,4,4] [0,1,2,3,4,5,6])
Eval (IsInfixOf [2,4,4] [0,1,2,3,4,5,6]) :: Bool
= False

data Elem (b :: a) (c :: [a]) (d :: Bool)

Return True if an element is in a list.

See also FindIndex.

Example

Expand

>>> :kind! Eval (Elem 1 [1,2,3])
Eval (Elem 1 [1,2,3]) :: Bool
= True
>>> :kind! Eval (Elem 1 [2,3])
Eval (Elem 1 [2,3]) :: Bool
= False

data Lookup (a :: k) (c :: [(k, b)]) (d :: Maybe b)

Find an element associated with a key in an association list.

data Find (b :: a -> Exp Bool) (c :: [a]) (d :: Maybe a)

Find Just the first element satisfying a predicate, or evaluate to Nothing if no element satisfies the predicate.

Example

Expand

>>> :kind! Eval (Find (TyEq 0) [1,2,3])
Eval (Find (TyEq 0) [1,2,3]) :: Maybe Natural
= Nothing

>>> :kind! Eval (Find (TyEq 0) [1,2,3,0])
Eval (Find (TyEq 0) [1,2,3,0]) :: Maybe Natural
= Just 0

data Filter (b :: a -> Exp Bool) (c :: [a]) (d :: [a])

Keep all elements that satisfy a predicate, remove all that don't.

Example

Expand

>>> :kind! Eval (Filter ((>) 3) [1,2,3,0])
Eval (Filter ((>) 3) [1,2,3,0]) :: [Natural]
= [1, 2, 0]

data Partition (b :: a -> Exp Bool) (c :: [a]) (d :: ([a], [a]))

Split a list into one where all elements satisfy a predicate, and a second where no elements satisfy it.

Example

Expand

>>> :kind! Eval (Partition ((>=) 35) [20, 30, 40, 50])
Eval (Partition ((>=) 35) [20, 30, 40, 50]) :: ([Natural],
                                                [Natural])
= '([20, 30], [40, 50])

data FindIndex (b :: a -> Exp Bool) (c :: [a]) (d :: Maybe Nat)

Find the index of an element satisfying the predicate.

Example

Expand

>>> :kind! Eval (FindIndex ((<=) 3) [1,2,3,1,2,3])
Eval (FindIndex ((<=) 3) [1,2,3,1,2,3]) :: Maybe Natural
= Just 2

>>> :kind! Eval (FindIndex ((>) 0) [1,2,3,1,2,3])
Eval (FindIndex ((>) 0) [1,2,3,1,2,3]) :: Maybe Natural
= Nothing

data SetIndex (b :: Nat) (c :: a) (d :: [a]) (e :: [a])

Modify an element at a given index.

The list is unchanged if the index is out of bounds.

Example

Expand

>>> :kind! Eval (SetIndex 2 7 [1,2,3])
Eval (SetIndex 2 7 [1,2,3]) :: [Natural]
= [1, 2, 7]

data ZipWith (d :: a -> b -> Exp c) (e :: [a]) (f :: [b]) (g :: [c])

Combine elements of two lists pairwise.

Example

Expand

>>> :kind! Eval (ZipWith (+) [1,2,3] [1,1,1])
Eval (ZipWith (+) [1,2,3] [1,1,1]) :: [Natural]
= [2, 3, 4]

data Zip (c :: [a]) (d :: [b]) (e :: [(a, b)])

data Unzip (c :: Exp [(a, b)]) (d :: ([a], [b]))

Bool

data UnBool (b :: Exp a) (c :: Exp a) (d :: Bool) (e :: a)

N.B.: The order of the two branches is the opposite of "if": UnBool ifFalse ifTrue bool.

This mirrors the default order of constructors:

data Bool = False | True
----------- False < True

data ((a :: Bool) || (b :: Bool)) (c :: Bool) infixr 2

data ((a :: Bool) && (b :: Bool)) (c :: Bool) infixr 3

data Not (a :: Bool) (b :: Bool)

Case splitting

data Case (a :: [Match j k]) (b :: j) (c :: k)

(Limited) equivalent of \case { .. } syntax. Supports matching of exact values (-->) and final matches for any value (Any) or for passing value to subcomputation (Else). Examples:

type BoolToNat = Case
  [ 'True  --> 0
  , 'False --> 1
  ]

type NatToBool = Case
  [ 0 --> 'False
  , Any   'True
  ]

type ZeroOneOrSucc = Case
  [ 0  --> 0
  , 1  --> 1
  , Else   ((+) 1)
  ]

data Match j k

type (-->) = 'Match_ :: j -> k -> Match j k infix 0

Match concrete type in Case.

type Is = 'Is_ :: (j -> Exp Bool) -> k -> Match j k

Match on predicate being successful with type in Case.

type Any = 'Any_ :: k -> Match j k

Match any type in Case. Should be used as a final branch.

Note: this identifier conflicts with Any (from Fcf.Class.Foldable) Any (from Data.Monoid), and Any (from GHC.Exts).

We recommend importing this one qualified.

type Else = 'Else_ :: (j -> Exp k) -> Match j k

Pass type being matched in Case to subcomputation. Should be used as a final branch.

Nat

data ((a :: Nat) + (b :: Nat)) (c :: Nat)

data ((a :: Nat) - (b :: Nat)) (c :: Nat)

data ((a :: Nat) * (b :: Nat)) (c :: Nat)

data ((a :: Nat) ^ (b :: Nat)) (c :: Nat)

data ((a :: Nat) <= (b :: Nat)) (c :: Bool)

data ((a :: Nat) >= (b :: Nat)) (c :: Bool)

data ((a :: Nat) < (b :: Nat)) (c :: Bool)

data ((a :: Nat) > (b :: Nat)) (c :: Bool)

Overloaded operations

data Map (c :: a -> Exp b) (d :: f a) (e :: f b)

Type-level fmap for type-level functors.

Note: this name clashes with Map from containers. FMap is provided as a synonym to avoid this.

Example

Expand

>>> data Example where Ex :: a -> Example  -- Hide the type of examples to avoid brittleness in different GHC versions
>>> data AddMul :: Nat -> Nat -> Exp Nat
>>> type instance Eval (AddMul x y) = (x TL.+ y) TL.* (x TL.+ y)
>>> :kind! Ex (Eval (Map (AddMul 2) '[0, 1, 2, 3, 4]) :: [Nat])
Ex (Eval (Map (AddMul 2) '[0, 1, 2, 3, 4]) :: [Nat]) :: Example
= Ex [4, 9, 16, 25, 36]

data Bimap (c :: a -> Exp a') (d :: b -> Exp b') (e :: f a b) (g :: f a' b')

Type-level bimap.

Example

Expand

>>> data Example where Ex :: a -> Example  -- Hide the type of examples to avoid brittleness in different GHC versions
>>> :kind! Ex (Eval (Bimap ((+) 1) (Flip (-) 1) '(2, 4)) :: (Natural, Natural))
Ex (Eval (Bimap ((+) 1) (Flip (-) 1) '(2, 4)) :: (Natural, Natural)) :: Example
= Ex '(3, 3)

Miscellaneous

data Error (b :: Symbol) (c :: a)

Type-level error.

data Constraints (a :: [Constraint]) b

Conjunction of a list of constraints.

data TyEq (c :: a) (d :: b) (e :: Bool)

Type equality.

Details

Expand

The base library also defines a similar (==); it differs from TyEq in the following ways:

TyEq is heterogeneous: its arguments may have different kinds;

TyEq is reflexive: TyEq a a always reduces to True even if a is a variable.

type family Stuck :: a

A stuck type that can be used like a type-level undefined.

class IsBool (b :: Bool) where

Methods

_If :: (b ~ 'True => r) -> (b ~ 'False => r) -> r

type family If (cond :: Bool) (tru :: k) (fls :: k) :: k where ...

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If 'True (tru :: k) (fls :: k) = tru

If 'False (tru :: k) (fls :: k) = fls

data Assert (a :: ErrorMessage) (b :: Exp Bool) (c :: r) (d :: r)

A compile-time assert.

Raises the provided TypeError, whenever the condition evaluates to False.

Usage example: type ExampleAssertionFailure = Eval ( Pure '["foo", "bar"] >>= Length >>= Assert ('Text Assertion) (TyEq Int Void) )

data AssertNot (a :: ErrorMessage) (b :: Exp Bool) (c :: r) (d :: r)

Compile-time assert, with condition negated.

Raises the provided TypeError, whenever the condition evaluates to True.

Also see Assert.