Fcf.Class.Ord

Description

Equality and ordering.

Note that equality doesn't really require a class, it can be defined uniformly as TyEq.

Order

data Compare (b :: a) (c :: a) (d :: Ordering)

Type-level compare for totally ordered data types.

Example

Expand

>>> :kind! Eval (Compare "a" "b")
Eval (Compare "a" "b") :: Ordering
= LT

>>> :kind! Eval (Compare '[1, 2, 3] '[1, 2, 3])
Eval (Compare '[1, 2, 3] '[1, 2, 3]) :: Ordering
= EQ

>>> :kind! Eval (Compare '[1, 3] '[1, 2])
Eval (Compare '[1, 3] '[1, 2]) :: Ordering
= GT

data ((b :: a) <= (c :: a)) (d :: Bool)

"Smaller than or equal to". Type-level version of (<=).

Example

Expand

>>> :kind! Eval ("b" <= "a")
Eval ("b" <= "a") :: Bool
= False

data ((b :: a) >= (c :: a)) (d :: Bool)

"Greater than or equal to". Type-level version of (>=).

Example

Expand

>>> :kind! Eval ("b" >= "a")
Eval ("b" >= "a") :: Bool
= True

data ((b :: a) < (c :: a)) (d :: Bool)

"Smaller than". Type-level version of (<).

Example

Expand

>>> :kind! Eval ("a" < "b")
Eval ("a" < "b") :: Bool
= True

data ((b :: a) > (c :: a)) (d :: Bool)

"Greater than". Type-level version of (>).

Example

Expand

>>> :kind! Eval ("b" > "a")
Eval ("b" > "a") :: Bool
= True

Equality

data TyEq (c :: a) (d :: b) (e :: Bool)

Type equality.

Details

Expand

The base library also defines a similar (==); it differs from TyEq in the following ways:

TyEq is heterogeneous: its arguments may have different kinds;

TyEq is reflexive: TyEq a a always reduces to True even if a is a variable.