Fcf

Description

First-class type families

For example, here is a regular type family:

type family   FromMaybe (a :: k) (m :: Maybe k) :: k
type instance FromMaybe a 'Nothing  = a
type instance FromMaybe a ('Just b) = b

With Fcf, it translates to a data declaration:

data FromMaybe :: k -> Maybe k -> Exp k
type instance Eval (FromMaybe a 'Nothing)  = a
type instance Eval (FromMaybe a ('Just b)) = b

Fcfs can be higher-order.

The kind constructor Exp is a monad: there's (=<<) and Pure.

Essential language extensions for Fcf:

{-# LANGUAGE
    DataKinds,
    PolyKinds,
    TypeFamilies,
    TypeOperators,
    UndecidableInstances #-}

First-class type families

type Exp a = a -> Type

Kind of type-level expressions indexed by their result type.

type family Eval (e :: Exp a) :: a

Expression evaluator.

type (@@) (f :: k1 -> Exp k) (x :: k1) = Eval (f x)

Apply and evaluate a unary type function.

Functional combinators

data Pure (b :: a) (c :: a)

data Pure1 (c :: a -> b) (d :: a) (e :: b)

data Pure2 (d :: a -> b -> c) (e :: a) (f :: b) (g :: c)

data Pure3 (e :: a -> b -> c -> d) (f :: a) (g :: b) (h :: c) (i :: d)

data ((c :: a -> Exp b) =<< (d :: Exp a)) (e :: b) infixr 1

data ((d :: b -> Exp c) <=< (e :: a -> Exp b)) (f :: a) (g :: c) infixr 1

type LiftM = (=<<) :: (a -> Exp b) -> Exp a -> b -> Type

data LiftM2 (d :: a -> b -> Exp c) (e :: Exp a) (f :: Exp b) (g :: c)

data LiftM3 (e :: a -> b -> c -> Exp d) (f :: Exp a) (g :: Exp b) (h :: Exp c) (i :: d)

data Join (b :: Exp (Exp a)) (c :: a)

data ((c :: a -> b) <$> (d :: Exp a)) (e :: b) infixl 4

data ((c :: Exp (a -> b)) <*> (d :: Exp a)) (e :: b) infixl 4

data Flip (d :: a -> b -> Exp c) (e :: b) (f :: a) (g :: c)

data ConstFn (c :: a) (d :: b) (e :: a)

data ((c :: a -> Exp b) $ (d :: a)) (e :: b) infixr 0

Note that this denotes the identity function, so ($) f can usually be replaced with f.

Operations on common types

Pairs

data Uncurry (d :: a -> b -> Exp c) (e :: (a, b)) (f :: c)

data Fst (c :: (a, b)) (d :: a)

data Snd (c :: (a, b)) (d :: b)

data ((a :: b -> Exp c) *** (d :: b' -> Exp c')) (e :: (b, b')) (f :: (c, c')) infixr 3

Specialization of Bimap for pairs.

Either

data UnEither (d :: a -> Exp c) (e :: b -> Exp c) (f :: Either a b) (g :: c)

data IsLeft (c :: Either a b) (d :: Bool)

data IsRight (c :: Either a b) (d :: Bool)

Maybe

data UnMaybe (c :: Exp b) (d :: a -> Exp b) (e :: Maybe a) (f :: b)

data FromMaybe (a :: k) (b :: Maybe k) (c :: k)

data IsNothing (b :: Maybe a) (c :: Bool)

data IsJust (b :: Maybe a) (c :: Bool)

Lists

data Foldr (c :: a -> b -> Exp b) (d :: b) (e :: t a) (f :: b)

Right fold.

Example

Expand

>>> :kind! Eval (Foldr (+) 0 [1, 2, 3, 4])
Eval (Foldr (+) 0 [1, 2, 3, 4]) :: Natural
= 10

data UnList (c :: b) (d :: a -> b -> Exp b) (e :: [a]) (f :: b)

This is Foldr with its argument flipped.

data ((b :: [a]) ++ (c :: [a])) (d :: [a])

List catenation.

Example

Expand

>>> data Example where Ex :: a -> Example  -- Hide the type of examples to avoid brittleness in different GHC versions
>>> :kind! Ex (Eval ([1, 2] ++ [3, 4]) :: [Natural])
Ex (Eval ([1, 2] ++ [3, 4]) :: [Natural]) :: Example
= Ex [1, 2, 3, 4]

data Filter (b :: a -> Exp Bool) (c :: [a]) (d :: [a])

Keep all elements that satisfy a predicate, remove all that don't.

Example

Expand

>>> :kind! Eval (Filter ((>) 3) [1,2,3,0])
Eval (Filter ((>) 3) [1,2,3,0]) :: [Natural]
= [1, 2, 0]

data Head (b :: [a]) (c :: Maybe a)

data Tail (b :: [a]) (c :: Maybe [a])

data Null (b :: [a]) (c :: Bool)

data Length (b :: [a]) (c :: Nat)

data Find (b :: a -> Exp Bool) (c :: [a]) (d :: Maybe a)

Find Just the first element satisfying a predicate, or evaluate to Nothing if no element satisfies the predicate.

Example

Expand

>>> :kind! Eval (Find (TyEq 0) [1,2,3])
Eval (Find (TyEq 0) [1,2,3]) :: Maybe Natural
= Nothing

>>> :kind! Eval (Find (TyEq 0) [1,2,3,0])
Eval (Find (TyEq 0) [1,2,3,0]) :: Maybe Natural
= Just 0

data FindIndex (b :: a -> Exp Bool) (c :: [a]) (d :: Maybe Nat)

Find the index of an element satisfying the predicate.

Example

Expand

>>> :kind! Eval (FindIndex ((<=) 3) [1,2,3,1,2,3])
Eval (FindIndex ((<=) 3) [1,2,3,1,2,3]) :: Maybe Natural
= Just 2

>>> :kind! Eval (FindIndex ((>) 0) [1,2,3,1,2,3])
Eval (FindIndex ((>) 0) [1,2,3,1,2,3]) :: Maybe Natural
= Nothing

data Lookup (a :: k) (c :: [(k, b)]) (d :: Maybe b)

Find an element associated with a key in an association list.

data SetIndex (b :: Nat) (c :: a) (d :: [a]) (e :: [a])

Modify an element at a given index.

The list is unchanged if the index is out of bounds.

Example

Expand

>>> :kind! Eval (SetIndex 2 7 [1,2,3])
Eval (SetIndex 2 7 [1,2,3]) :: [Natural]
= [1, 2, 7]

data ZipWith (d :: a -> b -> Exp c) (e :: [a]) (f :: [b]) (g :: [c])

Combine elements of two lists pairwise.

Example

Expand

>>> :kind! Eval (ZipWith (+) [1,2,3] [1,1,1])
Eval (ZipWith (+) [1,2,3] [1,1,1]) :: [Natural]
= [2, 3, 4]

data Zip (c :: [a]) (d :: [b]) (e :: [(a, b)])

data Unzip (c :: Exp [(a, b)]) (d :: ([a], [b]))

data Cons2 (c :: (a, b)) (d :: ([a], [b])) (e :: ([a], [b]))

Append elements to two lists. Used in the definition of Unzip.

Bool

data UnBool (b :: Exp a) (c :: Exp a) (d :: Bool) (e :: a)

N.B.: The order of the two branches is the opposite of "if": UnBool ifFalse ifTrue bool.

This mirrors the default order of constructors:

data Bool = False | True
----------- False < True

data ((a :: Bool) || (b :: Bool)) (c :: Bool) infixr 2

data ((a :: Bool) && (b :: Bool)) (c :: Bool) infixr 3

data Not (a :: Bool) (b :: Bool)

Case splitting

data Case (a :: [Match j k]) (b :: j) (c :: k)

(Limited) equivalent of \case { .. } syntax. Supports matching of exact values (-->) and final matches for any value (Any) or for passing value to subcomputation (Else). Examples:

type BoolToNat = Case
  [ 'True  --> 0
  , 'False --> 1
  ]

type NatToBool = Case
  [ 0 --> 'False
  , Any   'True
  ]

type ZeroOneOrSucc = Case
  [ 0  --> 0
  , 1  --> 1
  , Else   ((+) 1)
  ]

data Match j k

type (-->) = 'Match_ :: j -> k -> Match j k infix 0

Match concrete type in Case.

type Is = 'Is_ :: (j -> Exp Bool) -> k -> Match j k

Match on predicate being successful with type in Case.

type Any = 'Any_ :: k -> Match j k

Match any type in Case. Should be used as a final branch.

Note: this identifier conflicts with Any (from Fcf.Class.Foldable) Any (from Data.Monoid), and Any (from GHC.Exts).

We recommend importing this one qualified.

type Else = 'Else_ :: (j -> Exp k) -> Match j k

Pass type being matched in Case to subcomputation. Should be used as a final branch.

Nat

data ((a :: Nat) + (b :: Nat)) (c :: Nat)

data ((a :: Nat) - (b :: Nat)) (c :: Nat)

data ((a :: Nat) * (b :: Nat)) (c :: Nat)

data ((a :: Nat) ^ (b :: Nat)) (c :: Nat)

data ((a :: Nat) <= (b :: Nat)) (c :: Bool)

data ((a :: Nat) >= (b :: Nat)) (c :: Bool)

data ((a :: Nat) < (b :: Nat)) (c :: Bool)

data ((a :: Nat) > (b :: Nat)) (c :: Bool)

Overloaded operations

data Map (c :: a -> Exp b) (d :: f a) (e :: f b)

Type-level fmap for type-level functors.

Note: this name clashes with Map from containers. FMap is provided as a synonym to avoid this.

Example

Expand

>>> data Example where Ex :: a -> Example  -- Hide the type of examples to avoid brittleness in different GHC versions
>>> data AddMul :: Nat -> Nat -> Exp Nat
>>> type instance Eval (AddMul x y) = (x TL.+ y) TL.* (x TL.+ y)
>>> :kind! Ex (Eval (Map (AddMul 2) '[0, 1, 2, 3, 4]) :: [Nat])
Ex (Eval (Map (AddMul 2) '[0, 1, 2, 3, 4]) :: [Nat]) :: Example
= Ex [4, 9, 16, 25, 36]

data Bimap (c :: a -> Exp a') (d :: b -> Exp b') (e :: f a b) (g :: f a' b')

Type-level bimap.

Example

Expand

>>> data Example where Ex :: a -> Example  -- Hide the type of examples to avoid brittleness in different GHC versions
>>> :kind! Ex (Eval (Bimap ((+) 1) (Flip (-) 1) '(2, 4)) :: (Natural, Natural))
Ex (Eval (Bimap ((+) 1) (Flip (-) 1) '(2, 4)) :: (Natural, Natural)) :: Example
= Ex '(3, 3)

Miscellaneous

data Error (b :: Symbol) (c :: a)

Type-level error.

data Constraints (a :: [Constraint]) b

Conjunction of a list of constraints.

data TyEq (c :: a) (d :: b) (e :: Bool)

Type equality.

Details

Expand

The base library also defines a similar (==); it differs from TyEq in the following ways:

TyEq is heterogeneous: its arguments may have different kinds;

TyEq is reflexive: TyEq a a always reduces to True even if a is a variable.

type family Stuck :: a

A stuck type that can be used like a type-level undefined.

class IsBool (b :: Bool) where

Methods

_If :: (b ~ 'True => r) -> (b ~ 'False => r) -> r

type family If (cond :: Bool) (tru :: k) (fls :: k) :: k where ...

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If 'True (tru :: k) (fls :: k) = tru

If 'False (tru :: k) (fls :: k) = fls