Fcf

Description

First-class type families

For example, here is a regular type family:

type family   FromMaybe (a :: k) (m :: Maybe k) :: k
type instance FromMaybe a 'Nothing  = a
type instance FromMaybe a ('Just b) = b

With Fcf, it translates to a data declaration:

data FromMaybe :: k -> Maybe k -> Exp k
type instance Eval (FromMaybe a 'Nothing)  = a
type instance Eval (FromMaybe a ('Just b)) = b

Fcfs can be higher-order.

The kind constructor Exp is a monad: there's (=<<) and Pure.

Essential language extensions for Fcf:

{-# LANGUAGE
    DataKinds,
    PolyKinds,
    TypeFamilies,
    TypeInType,
    TypeOperators,
    UndecidableInstances #-}

First-class type families

type Exp a = a -> Type

Kind of type-level expressions indexed by their result type.

type family Eval (e :: Exp a) :: a

Expression evaluator.

type (@@) f x = Eval (f x)

Apply and evaluate a unary type function.

Functional combinators

data Pure :: a -> Exp a

data Pure1 :: (a -> b) -> a -> Exp b

data Pure2 :: (a -> b -> c) -> a -> b -> Exp c

data Pure3 :: (a -> b -> c -> d) -> a -> b -> c -> Exp d

data (=<<) :: (a -> Exp b) -> Exp a -> Exp b infixr 1

data (<=<) :: (b -> Exp c) -> (a -> Exp b) -> a -> Exp c infixr 1

type LiftM = (=<<)

data LiftM2 :: (a -> b -> Exp c) -> Exp a -> Exp b -> Exp c

data LiftM3 :: (a -> b -> c -> Exp d) -> Exp a -> Exp b -> Exp c -> Exp d

data Join :: Exp (Exp a) -> Exp a

data (<$>) :: (a -> b) -> Exp a -> Exp b infixl 4

data (<*>) :: Exp (a -> b) -> Exp a -> Exp b infixl 4

data Flip :: (a -> b -> Exp c) -> b -> a -> Exp c

data ConstFn :: a -> b -> Exp a

data ($) :: (a -> Exp b) -> a -> Exp b infixr 0

Note that this denotes the identity function, so ($) f can usually be replaced with f.

Operations on common types

Pairs

data Uncurry :: (a -> b -> Exp c) -> (a, b) -> Exp c

data Fst :: (a, b) -> Exp a

data Snd :: (a, b) -> Exp b

data (***) :: (b -> Exp c) -> (b' -> Exp c') -> (b, b') -> Exp (c, c') infixr 3

Equivalent to Bimap for pairs.

Either

data UnEither :: (a -> Exp c) -> (b -> Exp c) -> Either a b -> Exp c

data IsLeft :: Either a b -> Exp Bool

data IsRight :: Either a b -> Exp Bool

Maybe

data UnMaybe :: Exp b -> (a -> Exp b) -> Maybe a -> Exp b

data FromMaybe :: k -> Maybe k -> Exp k

data IsNothing :: Maybe a -> Exp Bool

data IsJust :: Maybe a -> Exp Bool

Lists

data Foldr :: (a -> b -> Exp b) -> b -> [a] -> Exp b

data UnList :: b -> (a -> b -> Exp b) -> [a] -> Exp b

N.B.: This is equivalent to a Foldr flipped.

data (++) :: [a] -> [a] -> Exp [a]

data Filter :: (a -> Exp Bool) -> [a] -> Exp [a]

data Head :: [a] -> Exp (Maybe a)

data Tail :: [a] -> Exp (Maybe [a])

data Null :: [a] -> Exp Bool

data Length :: [a] -> Exp Nat

data Find :: (a -> Exp Bool) -> [a] -> Exp (Maybe a)

data FindIndex :: (a -> Exp Bool) -> [a] -> Exp (Maybe Nat)

Find the index of an element satisfying the predicate.

type Lookup (a :: k) (as :: [(k, b)]) = (Map Snd (Eval (Find (TyEq a <=< Fst) as)) :: Exp (Maybe b))

Find an element associated with a key. Lookup :: k -> [(k, b)] -> Exp (Maybe b)

data SetIndex :: Nat -> a -> [a] -> Exp [a]

Modify an element at a given index.

The list is unchanged if the index is out of bounds.

data ZipWith :: (a -> b -> Exp c) -> [a] -> [b] -> Exp [c]

type Zip = ZipWith (Pure2 (,))

Zip :: [a] -> [b] -> Exp [(a, b)]

data Unzip :: Exp [(a, b)] -> Exp ([a], [b])

data Cons2 :: (a, b) -> ([a], [b]) -> Exp ([a], [b])

Bool

data UnBool :: Exp a -> Exp a -> Bool -> Exp a

N.B.: The order of the two branches is the opposite of "if": UnBool ifFalse ifTrue bool.

This mirrors the default order of constructors:

data Bool = False | True
----------- False < True

data (||) :: Bool -> Bool -> Exp Bool infixr 2

data (&&) :: Bool -> Bool -> Exp Bool infixr 3

data Not :: Bool -> Exp Bool

Multi-way if

data Guarded :: a -> [Guard (a -> Exp Bool) (Exp b)] -> Exp b

A conditional choosing the first branch whose guard a -> Exp Bool accepts a given value a.

Example

type UnitPrefix n = Eval (Guarded n
  '[ TyEq 0 ':= Pure ""
   , TyEq 1 ':= Pure "deci"
   , TyEq 2 ':= Pure "hecto"
   , TyEq 3 ':= Pure "kilo"
   , TyEq 6 ':= Pure "mega"
   , TyEq 9 ':= Pure "giga"
   , Otherwise ':= Error "Something else"
   ])

data Guard a b

A fancy-looking pair type to use with Guarded.

Constructors

a := b infixr 0

type Otherwise = ConstFn True

A catch-all guard for Guarded.

Nat

data (+) :: Nat -> Nat -> Exp Nat

data (-) :: Nat -> Nat -> Exp Nat

data * :: Nat -> Nat -> Exp Nat

data (^) :: Nat -> Nat -> Exp Nat

data (<=) :: Nat -> Nat -> Exp Bool

data (>=) :: Nat -> Nat -> Exp Bool

data (<) :: Nat -> Nat -> Exp Bool

data (>) :: Nat -> Nat -> Exp Bool

Overloaded operations

data Map :: (a -> Exp b) -> f a -> Exp (f b)

Type-level fmap for type-level functors.

data Bimap :: (a -> Exp a') -> (b -> Exp b') -> f a b -> Exp (f a' b')

Type-level bimap.

Miscellaneous

data Error :: Symbol -> Exp a

Type-level error.

data Constraints :: [Constraint] -> Exp Constraint

Conjunction of a list of constraints.

data TyEq :: a -> b -> Exp Bool

Type equality.

type family Stuck :: a

A stuck type that can be used like a type-level undefined.

class IsBool (b :: Bool) where

Methods

_If :: (b ~ True => r) -> (b ~ False => r) -> r

type family If (cond :: Bool) (tru :: k) (fls :: k) :: k where ...

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If True (tru :: k) (fls :: k) = tru

If False (tru :: k) (fls :: k) = fls