Fcf.Class.Bifunctor

Description

Bifunctors.

Bifunctors are "two-argument functors".

This module is the type-level equivalent of Data.Bifunctor.

Documentation

data Bimap (c :: a -> Exp a') (d :: b -> Exp b') (e :: f a b) (g :: f a' b')

Type-level bimap.

Example

Expand

>>> data Example where Ex :: a -> Example  -- Hide the type of examples to avoid brittleness in different GHC versions
>>> :kind! Ex (Eval (Bimap ((+) 1) (Flip (-) 1) '(2, 4)) :: (Natural, Natural))
Ex (Eval (Bimap ((+) 1) (Flip (-) 1) '(2, 4)) :: (Natural, Natural)) :: Example
= Ex '(3, 3)

data First (d :: a -> Exp b) (e :: f a c) (g :: f b c)

Type-level first. Apply a function along the first parameter of a bifunctor.

Example

Expand

>>> :kind! Eval (First ((+) 1) '(3,"a"))
Eval (First ((+) 1) '(3,"a")) :: (Natural, Symbol)
= '(4, "a")

data Second (b :: c -> Exp d) (e :: f a c) (g :: f a d)

Type-level second. Apply a function along the second parameter of a bifunctor.

This is generally equivalent to Map.

Example

Expand

>>> :kind! Eval (Second ((+) 1) '("a",3))
Eval (Second ((+) 1) '("a",3)) :: (Symbol, Natural)
= '("a", 4)