Fcf.Data.Bool

Description

Booleans.

Note that the operations from this module conflict with Data.Type.Bool.

Documentation

data UnBool (b :: Exp a) (c :: Exp a) (d :: Bool) (e :: a)

N.B.: The order of the two branches is the opposite of "if": UnBool ifFalse ifTrue bool.

This mirrors the default order of constructors:

data Bool = False | True
----------- False < True

data ((a :: Bool) || (b :: Bool)) (c :: Bool) infixr 2

data ((a :: Bool) && (b :: Bool)) (c :: Bool) infixr 3

data Not (a :: Bool) (b :: Bool)