Fcf.Data.Bool

Description

Booleans.

Note that the operations from this module conflict with Data.Type.Bool.

Documentation

data UnBool :: Exp a -> Exp a -> Bool -> Exp a

N.B.: The order of the two branches is the opposite of "if": UnBool ifFalse ifTrue bool.

This mirrors the default order of constructors:

data Bool = False | True
----------- False < True

data (||) :: Bool -> Bool -> Exp Bool infixr 2

data (&&) :: Bool -> Bool -> Exp Bool infixr 3

data Not :: Bool -> Exp Bool

Multi-way if

data Guarded :: a -> [Guard (a -> Exp Bool) (Exp b)] -> Exp b

A conditional choosing the first branch whose guard a -> Exp Bool accepts a given value a.

Example

type UnitPrefix n = Eval (Guarded n
  '[ TyEq 0 ':= Pure ""
   , TyEq 1 ':= Pure "deci"
   , TyEq 2 ':= Pure "hecto"
   , TyEq 3 ':= Pure "kilo"
   , TyEq 6 ':= Pure "mega"
   , TyEq 9 ':= Pure "giga"
   , Otherwise ':= Error "Something else"
   ])

data Guard a b

A fancy-looking pair type to use with Guarded.

Constructors

a := b infixr 0

type Otherwise = ConstFn True

A catch-all guard for Guarded.