Fcf.Data.List

Description

Lists.

See also Fcf.Class.Foldable for additional functions.

Basic functions

data ((b :: [a]) ++ (c :: [a])) (d :: [a])

List catenation.

Example

Expand

>>> data Example where Ex :: a -> Example  -- Hide the type of examples to avoid brittleness in different GHC versions
>>> :kind! Ex (Eval ([1, 2] ++ [3, 4]) :: [Natural])
Ex (Eval ([1, 2] ++ [3, 4]) :: [Natural]) :: Example
= Ex [1, 2, 3, 4]

data Head (b :: [a]) (c :: Maybe a)

data Last (b :: [a]) (c :: Maybe a)

data Tail (b :: [a]) (c :: Maybe [a])

data Cons (b :: a) (c :: [a]) (d :: [a])

Append an element to a list.

Example

Expand

>>> :kind! Eval (Cons 1 [2, 3])
Eval (Cons 1 [2, 3]) :: [Natural]
= [1, 2, 3]
>>> :kind! Eval (Cons Int [Char, Maybe Double])
Eval (Cons Int [Char, Maybe Double]) :: [*]
= [Int, Char, Maybe Double]

data Snoc (b :: [a]) (c :: a) (d :: [a])

Append an element to the end of a list.

Example

Expand

>>> :kind! Eval (Snoc [1,2,3] 4)
Eval (Snoc [1,2,3] 4) :: [Natural]
= [1, 2, 3, 4]

data Cons2 (c :: (a, b)) (d :: ([a], [b])) (e :: ([a], [b]))

Append elements to two lists. Used in the definition of Unzip.

data Init (b :: [a]) (c :: Maybe [a])

data Null (b :: [a]) (c :: Bool)

data Length (b :: [a]) (c :: Nat)

List transformations

data Reverse (b :: [a]) (c :: [a])

Reverse a list.

Example

Expand

>>> :kind! Eval (Reverse [1,2,3,4,5])
Eval (Reverse [1,2,3,4,5]) :: [Natural]
= [5, 4, 3, 2, 1]

data Intersperse (b :: a) (c :: [a]) (d :: [a])

Intersperse a separator between elements of a list.

Example

Expand

>>> :kind! Eval (Intersperse 0 [1,2,3,4])
Eval (Intersperse 0 [1,2,3,4]) :: [Natural]
= [1, 0, 2, 0, 3, 0, 4]

data Intercalate (b :: [a]) (c :: [[a]]) (d :: [a])

Join a list of words separated by some word.

Example

Expand

>>> :kind! Eval (Intercalate '[", "] [ '["Lorem"], '["ipsum"], '["dolor"] ])
Eval (Intercalate '[", "] [ '["Lorem"], '["ipsum"], '["dolor"] ]) :: [TL.Symbol]
= ["Lorem", ", ", "ipsum", ", ", "dolor"]

Reducing lists

See also Fcf.Class.Foldable.

data Foldr (c :: a -> b -> Exp b) (d :: b) (e :: t a) (f :: b)

Right fold.

Example

Expand

>>> :kind! Eval (Foldr (+) 0 [1, 2, 3, 4])
Eval (Foldr (+) 0 [1, 2, 3, 4]) :: Natural
= 10

data UnList (c :: b) (d :: a -> b -> Exp b) (e :: [a]) (f :: b)

This is Foldr with its argument flipped.

data Concat (a :: t m) (b :: m)

Concatenate a collection of elements from a monoid.

Example

Expand

For example, fold a list of lists.

Concat :: [[a]] -> Exp [a]

>>> :kind! Eval (Concat ([[1,2], [3,4], [5,6]]))
Eval (Concat ([[1,2], [3,4], [5,6]])) :: [Natural]
= [1, 2, 3, 4, 5, 6]
>>> :kind! Eval (Concat ([[Int, Maybe Int], [Maybe String, Either Double Int]]))
Eval (Concat ([[Int, Maybe Int], [Maybe String, Either Double Int]])) :: [*]
= [Int, Maybe Int, Maybe [Char], Either Double Int]

data ConcatMap (c :: a -> Exp [b]) (d :: t a) (e :: [b])

Map a function and concatenate the results.

This is FoldMap specialized to the list monoid.

Unfolding and building

data Unfoldr (c :: b -> Exp (Maybe (a, b))) (d :: b) (e :: [a])

Unfold a generator into a list.

Example

Expand

>>> data ToThree :: Nat -> Exp (Maybe (Nat, Nat))
>>> :{
type instance Eval (ToThree b) =
  If (4 TL.<=? b)
    Nothing
    (Just '(b, b TL.+ 1))
:}

>>> :kind! Eval (Unfoldr ToThree 0)
Eval (Unfoldr ToThree 0) :: [Natural]
= [0, 1, 2, 3]

See also the definition of Replicate.

data Replicate (b :: Nat) (c :: a) (d :: [a])

Repeat the same element in a list.

Example

Expand

>>> :kind! Eval (Replicate 4 '("ok", 2))
Eval (Replicate 4 '("ok", 2)) :: [(TL.Symbol, Natural)]
= ['("ok", 2), '("ok", 2), '("ok", 2), '("ok", 2)]

Sublists

data Take (b :: Nat) (c :: [a]) (d :: [a])

Take a prefix of fixed length.

Example

Expand

>>> :kind! Eval (Take 2 [1,2,3,4,5])
Eval (Take 2 [1,2,3,4,5]) :: [Natural]
= [1, 2]

data Drop (b :: Nat) (c :: [a]) (d :: [a])

Drop a prefix of fixed length, evaluate to the remaining suffix.

Example

Expand

>>> :kind! Eval (Drop 2 [1,2,3,4,5])
Eval (Drop 2 [1,2,3,4,5]) :: [Natural]
= [3, 4, 5]

data TakeWhile (b :: a -> Exp Bool) (c :: [a]) (d :: [a])

Take the longest prefix of elements satisfying a predicate.

Example

Expand

>>> :kind! Eval (TakeWhile ((>=) 3) [1, 2, 3, 4, 5])
Eval (TakeWhile ((>=) 3) [1, 2, 3, 4, 5]) :: [Natural]
= [1, 2, 3]

data DropWhile (b :: a -> Exp Bool) (c :: [a]) (d :: [a])

Drop the longest prefix of elements satisfying a predicate, evaluate to the remaining suffix.

Example

Expand

:kind! Eval (DropWhile ((>=) 3) [1, 2, 3, 4, 5]) Eval (DropWhile ((>=) 3) [1, 2, 3, 4, 5]) :: [Natural] = [4, 5]

data Span (b :: a -> Exp Bool) (c :: [a]) (d :: ([a], [a]))

Span, applied to a predicate p and a list xs, returns a tuple: the first component is the longest prefix (possibly empty) of xs whose elements satisfy p; the second component is the remainder of the list.

See also TakeWhile, DropWhile, and Break.

Example

Expand

>>> :kind! Eval (Span (Flip (<) 3) [1,2,3,4,1,2])
Eval (Span (Flip (<) 3) [1,2,3,4,1,2]) :: ([Natural], [Natural])
= '([1, 2], [3, 4, 1, 2])

>>> :kind! Eval (Span (Flip (<) 9) [1,2,3])
Eval (Span (Flip (<) 9) [1,2,3]) :: ([Natural], [Natural])
= '([1, 2, 3], '[])

>>> :kind! Eval (Span (Flip (<) 0) [1,2,3])
Eval (Span (Flip (<) 0) [1,2,3]) :: ([Natural], [Natural])
= '( '[], [1, 2, 3])

data Break (b :: a -> Exp Bool) (c :: [a]) (d :: ([a], [a]))

Break, applied to a predicate p and a list xs, returns a tuple: the first component is the longest prefix (possibly empty) of xs whose elements do not satisfy p; the second component is the remainder of the list.

Example

Expand

>>> :kind! Eval (Break (Flip (>) 3) [1,2,3,4,1,2])
Eval (Break (Flip (>) 3) [1,2,3,4,1,2]) :: ([Natural], [Natural])
= '([1, 2, 3], [4, 1, 2])

>>> :kind! Eval (Break (Flip (<) 9) [1,2,3])
Eval (Break (Flip (<) 9) [1,2,3]) :: ([Natural], [Natural])
= '( '[], [1, 2, 3])

>>> :kind! Eval (Break (Flip (>) 9) [1,2,3])
Eval (Break (Flip (>) 9) [1,2,3]) :: ([Natural], [Natural])
= '([1, 2, 3], '[])

data Tails (b :: [a]) (c :: [[a]])

List of suffixes of a list.

Example

Expand

>>> :kind! Eval (Tails [0,1,2,3])
Eval (Tails [0,1,2,3]) :: [[Natural]]
= [[0, 1, 2, 3], [1, 2, 3], [2, 3], '[3]]

Predicates

data IsPrefixOf (b :: [a]) (c :: [a]) (d :: Bool)

Return True when the first list is a prefix of the second.

Example

Expand

>>> :kind! Eval ([0,1,2] `IsPrefixOf` [0,1,2,3,4,5])
Eval ([0,1,2] `IsPrefixOf` [0,1,2,3,4,5]) :: Bool
= True

>>> :kind! Eval ([0,1,2] `IsPrefixOf` [0,1,3,2,4,5])
Eval ([0,1,2] `IsPrefixOf` [0,1,3,2,4,5]) :: Bool
= False

>>> :kind! Eval ('[] `IsPrefixOf` [0,1,3,2,4,5])
Eval ('[] `IsPrefixOf` [0,1,3,2,4,5]) :: Bool
= True

>>> :kind! Eval ([0,1,3,2,4,5] `IsPrefixOf` '[])
Eval ([0,1,3,2,4,5] `IsPrefixOf` '[]) :: Bool
= False

data IsSuffixOf (b :: [a]) (c :: [a]) (d :: Bool)

Return True when the first list is a suffix of the second.

Example

Expand

>>> :kind! Eval (IsSuffixOf [3,4,5] [0,1,2,3,4,5])
Eval (IsSuffixOf [3,4,5] [0,1,2,3,4,5]) :: Bool
= True

>>> :kind! Eval (IsSuffixOf [3,4,5] [0,1,3,2,4,5])
Eval (IsSuffixOf [3,4,5] [0,1,3,2,4,5]) :: Bool
= False

>>> :kind! Eval (IsSuffixOf '[] [0,1,3,2,4,5])
Eval (IsSuffixOf '[] [0,1,3,2,4,5]) :: Bool
= True

>>> :kind! Eval (IsSuffixOf [0,1,3,2,4,5] '[])
Eval (IsSuffixOf [0,1,3,2,4,5] '[]) :: Bool
= False

data IsInfixOf (b :: [a]) (c :: [a]) (d :: Bool)

Return True when the first list is contained within the second.

Example

Expand

>>> :kind! Eval (IsInfixOf [2,3,4] [0,1,2,3,4,5,6])
Eval (IsInfixOf [2,3,4] [0,1,2,3,4,5,6]) :: Bool
= True

>>> :kind! Eval (IsInfixOf [2,4,4] [0,1,2,3,4,5,6])
Eval (IsInfixOf [2,4,4] [0,1,2,3,4,5,6]) :: Bool
= False

Searching

data Elem (b :: a) (c :: [a]) (d :: Bool)

Return True if an element is in a list.

See also FindIndex.

Example

Expand

>>> :kind! Eval (Elem 1 [1,2,3])
Eval (Elem 1 [1,2,3]) :: Bool
= True
>>> :kind! Eval (Elem 1 [2,3])
Eval (Elem 1 [2,3]) :: Bool
= False

data Lookup (a :: k) (c :: [(k, b)]) (d :: Maybe b)

Find an element associated with a key in an association list.

data Find (b :: a -> Exp Bool) (c :: [a]) (d :: Maybe a)

Find Just the first element satisfying a predicate, or evaluate to Nothing if no element satisfies the predicate.

Example

Expand

>>> :kind! Eval (Find (TyEq 0) [1,2,3])
Eval (Find (TyEq 0) [1,2,3]) :: Maybe Natural
= Nothing

>>> :kind! Eval (Find (TyEq 0) [1,2,3,0])
Eval (Find (TyEq 0) [1,2,3,0]) :: Maybe Natural
= Just 0

data Filter (b :: a -> Exp Bool) (c :: [a]) (d :: [a])

Keep all elements that satisfy a predicate, remove all that don't.

Example

Expand

>>> :kind! Eval (Filter ((>) 3) [1,2,3,0])
Eval (Filter ((>) 3) [1,2,3,0]) :: [Natural]
= [1, 2, 0]

data Partition (b :: a -> Exp Bool) (c :: [a]) (d :: ([a], [a]))

Split a list into one where all elements satisfy a predicate, and a second where no elements satisfy it.

Example

Expand

>>> :kind! Eval (Partition ((>=) 35) [20, 30, 40, 50])
Eval (Partition ((>=) 35) [20, 30, 40, 50]) :: ([Natural],
                                                [Natural])
= '([20, 30], [40, 50])

Indexing lists

data FindIndex (b :: a -> Exp Bool) (c :: [a]) (d :: Maybe Nat)

Find the index of an element satisfying the predicate.

Example

Expand

>>> :kind! Eval (FindIndex ((<=) 3) [1,2,3,1,2,3])
Eval (FindIndex ((<=) 3) [1,2,3,1,2,3]) :: Maybe Natural
= Just 2

>>> :kind! Eval (FindIndex ((>) 0) [1,2,3,1,2,3])
Eval (FindIndex ((>) 0) [1,2,3,1,2,3]) :: Maybe Natural
= Nothing

data SetIndex (b :: Nat) (c :: a) (d :: [a]) (e :: [a])

Modify an element at a given index.

The list is unchanged if the index is out of bounds.

Example

Expand

>>> :kind! Eval (SetIndex 2 7 [1,2,3])
Eval (SetIndex 2 7 [1,2,3]) :: [Natural]
= [1, 2, 7]

Zipping and unzipping

data ZipWith (d :: a -> b -> Exp c) (e :: [a]) (f :: [b]) (g :: [c])

Combine elements of two lists pairwise.

Example

Expand

>>> :kind! Eval (ZipWith (+) [1,2,3] [1,1,1])
Eval (ZipWith (+) [1,2,3] [1,1,1]) :: [Natural]
= [2, 3, 4]

data Zip (c :: [a]) (d :: [b]) (e :: [(a, b)])

data Unzip (c :: Exp [(a, b)]) (d :: ([a], [b]))