Fcf.Data.Nat

Description

Natural numbers.

Note that the operators from this module conflict with GHC.TypeLits and GHC.TypeNats.

Reexported type

From GHC.TypeNats.

type Nat = Natural

A type synonym for Natural.

Previously, this was an opaque data type, but it was changed to a type synonym.

Since: base-4.16.0.0

Operations

data ((a :: Nat) + (b :: Nat)) (c :: Nat)

data ((a :: Nat) - (b :: Nat)) (c :: Nat)

data ((a :: Nat) * (b :: Nat)) (c :: Nat)

data ((a :: Nat) ^ (b :: Nat)) (c :: Nat)

Comparisons

Note that these conflict with Fcf.Class.Ord.

data ((a :: Nat) <= (b :: Nat)) (c :: Bool)

data ((a :: Nat) >= (b :: Nat)) (c :: Bool)

data ((a :: Nat) < (b :: Nat)) (c :: Bool)

data ((a :: Nat) > (b :: Nat)) (c :: Bool)