Fcf.Utils

Description

Miscellaneous families.

Documentation

data Error (b :: Symbol) (c :: a)

Type-level error.

data TError (b :: ErrorMessage) (c :: a)

TypeError as a fcf.

data Constraints (a :: [Constraint]) b

Conjunction of a list of constraints.

data TyEq (c :: a) (d :: b) (e :: Bool)

Type equality.

Details

Expand

The base library also defines a similar (==); it differs from TyEq in the following ways:

TyEq is heterogeneous: its arguments may have different kinds;

TyEq is reflexive: TyEq a a always reduces to True even if a is a variable.

type family Stuck :: a

A stuck type that can be used like a type-level undefined.

class IsBool (b :: Bool) where

Methods

_If :: (b ~ 'True => r) -> (b ~ 'False => r) -> r

data Case (a :: [Match j k]) (b :: j) (c :: k)

(Limited) equivalent of \case { .. } syntax. Supports matching of exact values (-->) and final matches for any value (Any) or for passing value to subcomputation (Else). Examples:

type BoolToNat = Case
  [ 'True  --> 0
  , 'False --> 1
  ]

type NatToBool = Case
  [ 0 --> 'False
  , Any   'True
  ]

type ZeroOneOrSucc = Case
  [ 0  --> 0
  , 1  --> 1
  , Else   ((+) 1)
  ]

data Match j k

type (-->) = 'Match_ :: j -> k -> Match j k infix 0

Match concrete type in Case.

type Is = 'Is_ :: (j -> Exp Bool) -> k -> Match j k

Match on predicate being successful with type in Case.

type Any = 'Any_ :: k -> Match j k

Match any type in Case. Should be used as a final branch.

Note: this identifier conflicts with Any (from Fcf.Class.Foldable) Any (from Data.Monoid), and Any (from GHC.Exts).

We recommend importing this one qualified.

type Else = 'Else_ :: (j -> Exp k) -> Match j k

Pass type being matched in Case to subcomputation. Should be used as a final branch.

From Data.Type.Bool

type family If (cond :: Bool) (tru :: k) (fls :: k) :: k where ...

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If 'True (tru :: k) (fls :: k) = tru

If 'False (tru :: k) (fls :: k) = fls